New "Go to console" and logout button and auto-redirect to the user profile page
now git won't store the .firebase directory anymore.
The user data observable has been improved, now can handle the user logout throwing an error.
Added the "Go to console" button in the user profile page and in the topbar.
Added also the logout button in the user profile page.
When the user logs he is redirected automagically to the userprofile page.
New interface for the user data stored in the firestore (user-data.ts)